+++ src/Providers/SiteInformationServiceProvider.php
@@ -36,26 +36,24 @@ class SiteInformationServiceProvider extends ServiceProvider
         $this->loadMigrationsFrom(__DIR__ . '/../../database/migrations');
         $this->loadViewsFrom(__DIR__ . '/../../resources/views', 'velor-site-information');
 
-        if ($config->get('velor-site-information.enabled') === true) {
-            $resources->register($this->configuredClass($config, 'velor-site-information.resources.site_information_subject', SiteInformationSubjectResource::class));
-            $resources->register($this->configuredClass($config, 'velor-site-information.resources.site_information', SiteInformationResource::class));
+        $resources->register($this->configuredClass($config, 'velor-site-information.resources.site_information_subject', SiteInformationSubjectResource::class));
+        $resources->register($this->configuredClass($config, 'velor-site-information.resources.site_information', SiteInformationResource::class));
 
-            $policies->register(
-                SiteInformationSubject::class,
-                $this->configuredClass($config, 'velor-site-information.policies.' . SiteInformationSubject::class, SiteInformationSubjectPolicy::class),
-            );
-            $policies->register(
-                SiteInformation::class,
-                $this->configuredClass($config, 'velor-site-information.policies.' . SiteInformation::class, SiteInformationPolicy::class),
-            );
+        $policies->register(
+            SiteInformationSubject::class,
+            $this->configuredClass($config, 'velor-site-information.policies.' . SiteInformationSubject::class, SiteInformationSubjectPolicy::class),
+        );
+        $policies->register(
+            SiteInformation::class,
+            $this->configuredClass($config, 'velor-site-information.policies.' . SiteInformation::class, SiteInformationPolicy::class),
+        );
 
-            $cmsMenuItems->registerBefore(
-                'users.index',
-                new CmsMenuItemData(SiteInformationSubject::class, 'site-information.show', 'velor-site-information::resources.site-information-subjects.plural', 'bi-info-circle'),
-            );
+        $cmsMenuItems->registerBefore(
+            'users.index',
+            new CmsMenuItemData(SiteInformationSubject::class, 'site-information.show', 'velor-site-information::resources.site-information-subjects.plural', 'bi-info-circle'),
+        );
 
-            $cmsRoutes->loadAuthenticated(__DIR__ . '/../../routes/cms.php');
-        }
+        $cmsRoutes->loadAuthenticated(__DIR__ . '/../../routes/cms.php');
 
         $this->publishes([
             __DIR__ . '/../../config/velor-site-information.php' => $this->app->configPath('velor-site-information.php'),
@@ -64,10 +62,6 @@ class SiteInformationServiceProvider extends ServiceProvider
         $this->publishes([
             __DIR__ . '/../../database/migrations' => $this->app->databasePath('migrations'),
         ], 'velor-site-information-migrations');
-
-        $this->publishes([
-            __DIR__ . '/../../database/seeders' => $this->app->databasePath('seeders'),
-        ], 'velor-site-information-seeders');
 
         $this->publishes([
             __DIR__ . '/../../lang' => $this->app->langPath('vendor/velor-site-information'),